[#4454 Phase 1E] Delete the codegen compiled-query path
Phase 1D (#4456) made the JasperFx.RuntimeCompiler-driven path runtime-
unreachable for live compiled-query dispatch — every session.Query(...)
now goes through CompiledQueryHandlerRegistry (source-gen-emitted
descriptor) or RuntimeCompiledQueryDescriptorFactory (reflection
fallback) and lands in SourceGeneratedCompiledQuerySource<TOut>. The
Roslyn-emit classes and the ICodeFile contract for compiled queries
were carrying their weight as a "just in case" PoC bridge; the bridge
is gone. Delete it.

Files removed:
 - CompiledQueryCodeFile.cs (ICodeFile producer for the codegen CLI)
 - CompiledQuerySourceBuilder.cs (Roslyn-emit assembly builder)
 - CompiledQuerySource.cs (codegen base for emitted sources)
 - CompiledSourceType.cs (the Stateless/Cloneable/Complex enum)
 - StatelessCompiledQuery.cs / ClonedCompiledQuery.cs / ComplexCompiledQuery.cs
   (the three codegen-emitted handler base classes — superseded by
   SourceGeneratedStatelessHandler / SourceGeneratedClonedHandler /
   SourceGeneratedComplexHandler)
 - WriteSerializedJsonParameterFrame.cs (codegen Frame for the
   containment / JsonPath JSON-payload writers; runtime equivalent is
   CompiledQueryDictionaryBuilder)

Member-level deletions:
 - ICompiledQueryAwareFilter.GenerateCode — the source-gen + FEC paths
   consume BuildSetter() now; no codegen consumer remains. All seven
   filter implementations drop their GenerateCode overrides:
     · StringContainsFilter, StringStartsWithFilter,
       StringEndsWithFilter, StringEqualsIgnoreCaseFilter
     · ContainmentWhereFilter, ChildCollectionJsonPathCountFilter,
       DictionaryContainsKeyFilter
 - ParameterUsage.GenerateCode / generateSimpleCode / generateEnumCode
   / npgsqlArrayDbTypeCodeFor — the runtime descriptor factory owns
   the same dispatch in RuntimeCompiledQueryDescriptorFactory.
 - CompiledQueryPlan.GenerateCode — orchestrated the codegen emit
   over the plan's commands; no caller.
 - DocumentStore.CompiledQueryCollection's ICodeFileCollection.BuildFiles()
   compiled-query branch — Phase 5 will retire the `dotnet marten
   codegen ...` CLI surface and the ICodeFileCollection contract along
   with it. For now BuildFiles returns Array.Empty so non-compiled-query
   ICodeFile consumers keep working.

The DictionaryValueUsage class survives the WriteSerializedJsonParameterFrame
deletion — moved to its own file because the runtime
CompiledQueryDictionaryBuilder + filter Apply paths still need it.
The codegen-only IDictionaryPart tree (DictionaryDeclaration,
ArrayContainer, ArrayDeclaration, ArrayScalarValue, DictionaryValue)
had no other consumers and is gone.

Verification:

    dotnet test src/CompiledQueryTests/CompiledQueryTests.csproj
    → 9/9 pass

    dotnet test src/LinqTests/LinqTests.csproj --filter "FullyQualifiedName~Compiled"
    → 75/75 pass

901 lines of dead code removed.
