--- src/Marten/Internal/CompiledQueries/CompiledSourceType.cs
@@ -1,8 +1,0 @@
-namespace Marten.Internal.CompiledQueries;
-
-internal enum CompiledSourceType
-{
-    Stateless,
-    Complex,
-    Cloneable
-}
