+++ src/Marten/Internal/CompiledQueries/IQueryMember.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
-using JasperFx.CodeGeneration;
 using Npgsql;
 
 namespace Marten.Internal.CompiledQueries;
