--- src/Marten/Internal/CompiledQueries/StatelessCompiledQuery.cs
@@ -1,57 +1,0 @@
-using System.Data.Common;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-using Marten.Linq.QueryHandlers;
-using Weasel.Postgresql;
-
-namespace Marten.Internal.CompiledQueries;
-
-public abstract class StatelessCompiledQuery<TOut, TQuery>: IQueryHandler<TOut>
-{
-    private readonly IQueryHandler<TOut> _inner;
-    protected readonly TQuery _query;
-
-    public StatelessCompiledQuery(IQueryHandler<TOut> inner, TQuery query)
-    {
-        _inner = inner;
-        _query = query;
-    }
-
-    public abstract void ConfigureCommand(ICommandBuilder builder, IMartenSession session);
-
-    public Task<int> StreamJson(Stream stream, DbDataReader reader, CancellationToken token)
-    {
-        return _inner.StreamJson(stream, reader, token);
-    }
-
-    public Task<TOut> HandleAsync(DbDataReader reader, IMartenSession session, CancellationToken token)
-    {
-        return _inner.HandleAsync(reader, session, token);
-    }
-
-    protected string StartsWith(string value)
-    {
-        return $"{EscapeLikeValue(value)}%";
-    }
-
-    protected string ContainsString(string value)
-    {
-        return $"%{EscapeLikeValue(value)}%";
-    }
-
-    protected string EndsWith(string value)
-    {
-        return $"%{EscapeLikeValue(value)}";
-    }
-
-    protected string EqualsIgnoreCaseValue(string value)
-    {
-        return EscapeLikeValue(value);
-    }
-
-    private static string EscapeLikeValue(string value)
-    {
-        return value.Replace("\\", "\\\\").Replace("%", "\\%").Replace("_", "\\_");
-    }
-}
