+++ src/Marten/Linq/Parsing/Methods/Strings/StringEquals.cs
@@ -1,7 +1,6 @@
 #nullable enable
 using System;
 using System.Reflection;
-using JasperFx.CodeGeneration;
 using JasperFx.Core.Reflection;
 using Marten.Internal.CompiledQueries;
 using Marten.Linq.Members;
@@ -62,16 +61,6 @@ internal class StringEqualsIgnoreCaseFilter : ISqlFragment, ICompiledQueryAwareF
         }
 
         return false;
-    }
-
-    public void GenerateCode(GeneratedMethod method, int parameterIndex, string parametersVariableName)
-    {
-        var maskedValue = $"EqualsIgnoreCaseValue(_query.{_queryMember!.Name})";
-
-        method.Frames.Code($@"
-{parametersVariableName}[{parameterIndex}].NpgsqlDbType = {{0}};
-{parametersVariableName}[{parameterIndex}].Value = {maskedValue};
-", NpgsqlDbType.Varchar);
     }
 
     public Action<NpgsqlParameter, object> BuildSetter()
